+++ frontend/src/components/employer/CreateProject.vue
@@ -236,8 +236,8 @@
 
 <script>
 import axios from 'axios'
-import MainEmployerHeader from './common/MainEmployerHeader.vue'
-import '../assets/neumorphismGlobal.css'
+import MainEmployerHeader from '../common/MainEmployerHeader.vue'
+import '../../assets/neumorphismGlobal.css'
 
 export default {
   // 1. Nombre del componente
