+++ frontend/src/components/employer/projectDashboard/AddBenefit.vue
@@ -240,8 +240,8 @@
 </template>
 
 <script>
-import MainEmployerHeader from '../common/MainEmployerHeader.vue'
-import { apiConfig } from '../../config/api.js'
+import MainEmployerHeader from '../../common/MainEmployerHeader.vue'
+import { apiConfig } from '../../../config/api.js'
 
 export default {
   name: 'AddBenefit',
