+++ frontend/src/components/employer/projectDashboard/EmployeesFilter.vue
@@ -52,7 +52,7 @@
 </template>
 
 <script>
-import "../assets/Neumorfismo.css";
+// import "../assets/Neumorfismo.css";
 
 export default {
   name: "EmployeesFilter",
