+++ frontend/src/components/employer/projectDashboard/EmployeesList.vue
@@ -81,7 +81,7 @@
 </template>
 
 <script>
-  import "../assets/Neumorfismo.css"
+  // import "../assets/Neumorfismo.css"
 
   export default {
     name: 'ListaEmpleados',
