+++ frontend/src/components/employer/projectDashboard/RegisterEmployee.vue
@@ -206,7 +206,7 @@
 </template>
 
 <script>
-import MainEmployerHeader from '../common/MainEmployerHeader.vue'
+import MainEmployerHeader from '../../common/MainEmployerHeader.vue'
 import DashboardProjectSubHeader from './DashboardProjectSubHeader.vue'
 
 export default {
