+++ frontend/src/main.js
@@ -12,19 +12,20 @@ import './assets/neumorphismGlobal.css'
 import './assets/tailwind.css'
 
 // Vue Components
+import DashboardEmployee from './components/employee/DashboardEmployee.vue'
 import LandingPage from './components/LandingPage.vue'
 import Login from './components/LoginPage.vue'
 import PasswordSetup from './components/PasswordSetup.vue'
-import DashboardMainEmployer from './components/DashboardMainEmployer.vue'
 import SuperAdminMenu from './components/SuperAdminMenu.vue'
-import CreateProject from './components/CreateProject.vue'
 import SignUpEmployer from './components/SignUpEmployer.vue'
 import HoursRegistry from './components/HoursRegistry.vue'
-import DashboardProject from './components/projectDashboard/DashboardProject.vue'
-import RegisterEmployee from './components/projectDashboard/RegisterEmployee.vue'
 import ProfileEmployee from './components/ProfileEmployee.vue'
-import DashboardEmployee from './components/employee/DashboardEmployee.vue'
-import AddBenefit from './components/projectDashboard/AddBenefit.vue'
+// EMPLOYER COMPONENTS
+import AddBenefit from './components/employer/projectDashboard/AddBenefit.vue'
+import RegisterEmployee from './components/employer/projectDashboard/RegisterEmployee.vue'
+import DashboardMainEmployer from './components/employer/DashboardMainEmployer.vue'
+import DashboardProject from './components/employer/projectDashboard/DashboardProject.vue'
+import CreateProject from './components/employer/CreateProject.vue'
 
 // Route definitions
 const routes = [
